vertigo-debug-toolkit/requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-SocketIO==5.3.4
python-socketio==5.8.0
Werkzeug==2.3.7

# Langfuse and LLM Integration
langfuse==3.2.1
langchain==0.0.267
langchain-openai==0.0.2
langchain-community==0.0.10
google-generativeai==0.8.5

# Database
SQLAlchemy==2.0.21
alembic==1.12.0

# HTTP and API
requests==2.31.0
httpx==0.24.1

# Background Task Scheduling
APScheduler==3.10.4

# Google Cloud
google-cloud-firestore==2.16.0
google-cloud-core==2.4.1
google-auth==2.29.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
sentence-transformers==5.0.0
scikit-learn==1.7.1

# Visualization
plotly==5.16.1

# Authentication and Security
bcrypt==4.0.1
python-dotenv==1.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
cryptography==41.0.4

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.2.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-xdist==3.3.1
pytest-html==3.2.0
factory-boy==3.3.0
freezegun==1.2.2

# Development
black==23.7.0
flake8==6.0.0
pre-commit==3.3.3

# Security and Validation
bleach==6.0.0
wtforms==3.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# System Monitoring (for performance metrics)
psutil==5.9.5

# Caching (optional but recommended)
redis==4.6.0 